implements protected goals routes

+++ backend/controllers/goalController.js
@@ -8,13 +8,14 @@ can use package express-async-handler
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 /* @desc    Get goals
    @route   Get /api/goals
    @access  Private 
  */
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({ user: req.user.id })
 
   res.status(200).json(goals)
 })
@@ -32,7 +33,8 @@ const setGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id
   })
 
   res.status(200).json(goal)
@@ -49,9 +51,21 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error('Goal not found')
   }
 
+  // Check for user
+  if (!req.user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Check logged in user matches goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true
-  }) // new option creates if doesn't exists
+  }) // new option returns updated data
 
   // console.log(updatedGoal)
   res.status(200).json(updatedGoal)
@@ -67,6 +81,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
+  }
+
+  const user = User.findById(req.user.id)
+  // Check for user
+  if (!req.user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Check logged in user matches goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await Goal.findByIdAndDelete(req.params.id)
